refactor(@angular-devkit/core): move existing workspace API under experimental
This API is already experimental.  The move allows for the new stable API to use the "workspace" directory structure and to also better signify that the existing API is in fact experimental and subject to change or removal.
