+++ packages/angular_devkit/core/src/experimental/workspace/test/test-workspace.json
